Create an infinite scroll of posts
In this change, I created a Posts.tsx file that can allow infinite scroll of posts with loading when navigatimg down the posts.
